LayoutTests/http/tests/security/contentSecurityPolicy/injected-inline-style-blocked.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Security-Policy" content="style-src 'none'">
<script src="resources/dump-as-text.js"></script>
</head>
<body>
<div id="test1">
  PASS 1/2
</div>
<div id="test2">
  PASS 2/2
</div>
<script src="resources/inject-style.js"></script>
<script>
document.write(document.styleSheets.length);
</script>
</body>
</html>
